feat: add locale component
